requirements.txt
tk
requests
python-dotenv
qrcode
reportlab
openpyxl
flask          # for dummy_server only
